MICRO-221: add code as a required argument for reset password
This is provided by the BC mutation requestPasswordResetEmail and will be returned as a query paramater named 'c'. As AC uses email rather than this code, more investigation is required to get this to work for AC. Front end changes are also required to ensure we are sending this argument through.